+++ src/main/java/jnt/scimark2/MonteCarlo.java
@@ -31,7 +31,7 @@ package jnt.scimark2;
 public class MonteCarlo {
     final static int SEED = 113;
 
-    public static final double num_flops(int Num_samples) {
+    public static final double num_flops(long Num_samples) {
         // 3 flops in x^2+y^2 and 1 flop in random routine
 
         return ((double) Num_samples) * 4.0;
@@ -39,13 +39,13 @@ public class MonteCarlo {
     }
 
 
-    public static final double integrate(int Num_samples) {
+    public static final double integrate(long Num_samples) {
 
         Random R = new Random(SEED);
 
 
-        int under_curve = 0;
-        for (int count = 0; count < Num_samples; count++) {
+        long under_curve = 0;
+        for (long count = 0; count < Num_samples; count++) {
             double x = R.nextDouble();
             double y = R.nextDouble();
